feat/#130: renamed objects for destinction
