+++ src/components/hooks/useAuthState.js
@@ -1,23 +1,23 @@
 import React, { useState, useEffect, useContext, createContext } from "react";
-import auth from "../services/authentication";
+import auth from "../services/authService";
 
-const userContext = createContext();
+const authStateContext = createContext();
 
 // Provider component that wraps your app and makes user object ...
 // ... available to any child component that calls useUser().
-export function UserProvider({ children }) {
-  const [user, setUser] = useState({ user: "initialising" });
+export function AuthStateProvider({ children }) {
+  const [userId, setUserId] = useState("initialising");
 
   // Subscribe to firebase user state changes on mount
   // Because this sets state in the callback it will cause any
   // component that utilizes this hook to re-render with the
   // latest user object.
   useEffect(() => {
-    const unsubscribe = auth.onAuthStateChanged((user) => {
-      if (user) {
-        setUser(user);
+    const unsubscribe = auth.onAuthStateChanged((userId) => {
+      if (userId) {
+        setUserId(userId);
       } else {
-        setUser(null);
+        setUserId(null);
       }
     });
 
@@ -25,10 +25,14 @@ export function UserProvider({ children }) {
     return () => unsubscribe();
   }, []);
 
-  return <userContext.Provider value={user}>{children}</userContext.Provider>;
+  return (
+    <authStateContext.Provider value={userId}>
+      {children}
+    </authStateContext.Provider>
+  );
 }
 
 // Hook for child components to get the user object and re-render when it changes.
-export const useUser = () => {
-  return useContext(userContext);
+export const useAuthState = () => {
+  return useContext(authStateContext);
 };
